+++ datadog_checks_base/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - datadog_checks_base
 
+## 25.0.1 / 2022-02-24
+
+* [Fixed] Properly raise scraper error in OpenMetrics v2. See [#11564](https://github.com/DataDog/integrations-core/pull/11564).
+
 ## 25.0.0 / 2022-02-19
 
 * [Added] Add `pyproject.toml` file. See [#11301](https://github.com/DataDog/integrations-core/pull/11301).
@@ -17,7 +21,7 @@
 * [Added] Upgrade psutil to 5.9.0. See [#11139](https://github.com/DataDog/integrations-core/pull/11139).
 * [Changed] Add tls_protocols_allowed configuration option. See [#11237](https://github.com/DataDog/integrations-core/pull/11237).
 
-## 23.7.6 / 2022-02-03
+## 23.7.6 / 2022-02-03 / Agent 7.34.0
 
 * [Fixed] Update obfuscator wrapper to return empty string. See [#11277](https://github.com/DataDog/integrations-core/pull/11277).