+++ kube_metrics_server/CHANGELOG.md
@@ -6,7 +6,7 @@
 * [Fixed] Fix namespace packaging on Python 2. See [#11532](https://github.com/DataDog/integrations-core/pull/11532).
 * [Changed] Add tls_protocols_allowed option documentation. See [#11251](https://github.com/DataDog/integrations-core/pull/11251).
 
-## 1.4.1 / 2022-01-18
+## 1.4.1 / 2022-01-18 / Agent 7.34.0
 
 * [Fixed] Fix the type of `bearer_token_auth`. See [#11144](https://github.com/DataDog/integrations-core/pull/11144).