+++ kube_proxy/CHANGELOG.md
@@ -6,7 +6,7 @@
 * [Fixed] Fix namespace packaging on Python 2. See [#11532](https://github.com/DataDog/integrations-core/pull/11532).
 * [Changed] Add tls_protocols_allowed option documentation. See [#11251](https://github.com/DataDog/integrations-core/pull/11251).
 
-## 4.0.2 / 2022-01-21
+## 4.0.2 / 2022-01-21 / Agent 7.34.0
 
 * [Fixed] Remove new kube proxy metric with high cardinality. See [#11182](https://github.com/DataDog/integrations-core/pull/11182).