+++ silk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG - Silk
 
+## 1.0.2 / 2022-03-01
+
+* [Fixed] Update system state log line wording . See [#11577](https://github.com/DataDog/integrations-core/pull/11577).
+* [Fixed] Update example port. See [#11578](https://github.com/DataDog/integrations-core/pull/11578).
+
+## 1.0.1 / 2022-02-23
+
+* [Fixed] Bump base package dependency. See [#11549](https://github.com/DataDog/integrations-core/pull/11549).
+
 ## 1.0.0 / 2022-02-19
 
 * [Added] Add Silk integration. See [#10783](https://github.com/DataDog/integrations-core/pull/10783).